Allow user to verify email if verification is required
Previously the lambda for requesting and confirming verification
required is set to be `user_required=True`. If the server has a setting
to check for require verification when calling API, the client will not
be able to verify their user data.

This fixes the problem by manually checking if there is a user currently
logged in.

refs #42